// types/events.ts
export interface EventRecord {
  id: string;
  sourceRow: string;
  year: string;
  entryId: string;
  clientName: string;
  eventDate: string;
  venueName: string;
  eventType: string;
  contactPhone: string;
  email: string;
  eventAddress: string;
  estGuestCount: string;
  numberOfArtists: string;
  artistNames: string;
  counterStaffCharge: string;
  counterNames: string;
  tempFacilityLicenseFee: string;
  customFlash: string;
  customFlashFee: string;
  temporaryTattoos: string;
  temporaryTattooFee: string;
  travelDistance: string;
  radiusFee: string;
  setupTime: string;
  eventStartTime: string;
  eventEndTime: string;
  extraHours: string;
  extraHourlyCharge: string;
  optionalFee: string;
  depositRequired: string;
  depositDueDate: string;
  balanceDueDate: string;
  contractUrl: string;
  tflUrl: string;
  artImageUrl: string;
  payStatus: string;
  contractNotes: string;
  privateNotes: string;
  status: string;
  artistFeeBreakdown: string;
  gravitySubmissionRaw: string;
  gravityImportedAt: string;
  eventCompletedAt: string;
  autoPricingEnabled: string;
  staffPriceAdjustment: string;
  staffPriceAdjustmentReason: string;
  creditApplied: string;
  collectFullPayout: string;
  dontUseCredit: string;
}

export const EMPTY_EVENT: EventRecord = {
  id: '',
  sourceRow: '',
  year: '2026',
  entryId: '',
  clientName: '',
  eventDate: '',
  venueName: '',
  eventType: '',
  contactPhone: '',
  email: '',
  eventAddress: '',
  estGuestCount: '',
  numberOfArtists: '',
  artistNames: '',
  counterStaffCharge: '',
  counterNames: '',
  tempFacilityLicenseFee: '200',
  customFlash: 'NO',
  customFlashFee: '',
  temporaryTattoos: 'NO',
  temporaryTattooFee: '',
  travelDistance: '',
  radiusFee: '',
  setupTime: '',
  eventStartTime: '',
  eventEndTime: '',
  extraHours: '',
  extraHourlyCharge: '',
  optionalFee: '',
  depositRequired: '',
  depositDueDate: '',
  balanceDueDate: '',
  contractUrl: '',
  tflUrl: '',
  artImageUrl: '',
  payStatus: '',
  contractNotes: '',
  privateNotes: '',
  status: '',
  artistFeeBreakdown: '',
  gravitySubmissionRaw: '',
  gravityImportedAt: '',
  eventCompletedAt: '',
  autoPricingEnabled: 'YES',
  staffPriceAdjustment: '',
  staffPriceAdjustmentReason: '',
  creditApplied: '',
  collectFullPayout: 'NO',
  dontUseCredit: 'NO',
};